requirements.txt
snowflake-connector-python
pandas~=3.0.1
python-dotenv~=1.2.2